requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
requests>=2.31.0
beautifulsoup4>=4.12.3
lxml>=5.0.0
fastembed>=0.3.0
faiss-cpu>=1.8.0
groq>=0.9.0
python-dotenv>=1.0.0
numpy>=1.26.0
